% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{aggregate_df}
\alias{aggregate_df}
\title{Compute summary statistics within groups in a nested data frame}
\usage{
aggregate_df(data, id, remove_var = NULL, prefix_out = NULL,
  intake_var = NULL, out_values = c("mean", "sd", "count", "sum", "missing",
  "cor", "min", "max", "true"))
}
\arguments{
\item{data}{The data frame in long-format that contains the variables to be analysed}

\item{id}{Character string. The id or grouping variable, in which other observations are nested}

\item{remove_var}{Character string. Variables that should be removed before the computation proceeds (otherwise, the function assumes that all other variables should be used for the computation)}

\item{prefix_out}{The prefix for the variables that are returned. Default is \code{NULL}. If \code{NULL}, the input from \code{id} is used as prefix.}

\item{intake_var}{Character string. Specific variables (e.g. names of these variables) for which the ananylses should be run, i. e. if only a subset of variables should be used.}

\item{out_values}{The values to be returned. Can be either
\describe{
  \item{\code{"mean"}}{Computes the mean within each group/id. Missing values are removed before computation.}
  \item{\code{"sd"}}{Computes the sd each group/id. Missing values are removed before computation.}
  \item{\code{"count"}}{Computes the number of cases within each group/id, including missings.}
  \item{\code{"sum"}}{Computes the sum of values within each group/id. Missing values are removed before computation.}
  \item{\code{"missing"}}{Counts the number of missing values in each group/id.}
  \item{\code{"cor"}}{Computes the within-correlation for each variable within each group/id. Pearson correlation (\code{\link{cor}}) with \code{"pairwise.complete.obs"} is used.}
  \item{\code{"max"}}{Returns the maximum in each group/id, ignoring NAs.}
  \item{\code{"min"}}{Returns the minimum in each group/id, ignoring NAs.}
  \item{\code{"true"}}{Per group/id, returns 0 if the group/id contains any other value than 0, otherwise returns 1. Missings are ignored.}
}}
}
\value{
Returns a data frame in wide format (i. e. one row per group/id). Variable names are the original variable names with a correspondng prefix and an underscore (e. g. \code{mean_} for the mean). For the correlations, the names of the two variables that are correlated with each other are pasted together and the prefix \code{cor} is added, e. g. \code{cor.var_1.var_2} for the correlation between \code{var_1} and \code{var_2}.
}
\description{
This function allows the computation of some summary statistics within groups/persons or other clusters. The data needs to be in a long format. At least one variable needs to be a grouping-variable, e.g. \code{id}. Other variables of the data frame should be numeric.
}
\examples{
df <- aggregate_df(wide_example_data, id="id")
head(df)
}

