% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{check_exist_wd}
\alias{check_exist_wd}
\title{Check for existing objects in working directory}
\usage{
check_exist_wd(check = NULL, folder = NULL)
}
\arguments{
\item{check}{The object that is checked for in the current folder}

\item{folder}{The folder that is checked in. If NULL, then the current working directory is checked}
}
\value{
logical, does the tested object exist in the folder?
}
\description{
The function checks if an object exists in the current working direcotry or specified folder.
}

