% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{check_exist_ws}
\alias{check_exist_ws}
\title{Checks if an object exists in the global environment}
\usage{
check_exist_ws(test_object)
}
\arguments{
\item{test_object}{The object that is tested/checked}
}
\value{
logical, does the tested object exist in the workspace?
}
\description{
The global environment is the same as the workspace.
}
\examples{
a <- 1
check_exist_ws(a)
rm(a)
check_exist_ws(a)
}

