% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{fac_to_chr}
\alias{fac_to_chr}
\title{Turns factors to character strings in a data frame}
\usage{
fac_to_chr(data)
}
\arguments{
\item{data}{The data frame that should be converted}
}
\value{
A data frame in which all factors are converted to character strings.
}
\description{
Turns factors to character strings in a data frame
}
\examples{
data <- data.frame(a = c(1, 2, 3), b = as.factor(c(a, b, c)))
fac_to_chr(data)
}

