% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{gather_multiple}
\alias{gather_multiple}
\title{Gather multiple variables at the same time and turn them into long format}
\usage{
gather_multiple(data, key_vars, varying = NULL, search = "\\\\d$")
}
\arguments{
\item{data}{The data frame that contains the variables}

\item{key_vars}{The key or id variables that should be the grouping variables in the long format}

\item{varying}{The variables that are varying and should be returned in a long format. If NULL, all other variables except \code{key_vars} are taken and turned into long format. Default is to search for numbers at the end (\code{"\\d$"}}

\item{search}{The search string that is used to detect item-stems}
}
\value{
A data frame in long format, with the \code{key_vars} as grouping variables and the \code{varying} item-stems as stacked variables.
}
\description{
Gather multiple variables at the same time and turn them into long format
}
\examples{
gather_multiple(wide_example_data, key_vars = c("id", "id2"), varying = NULL)
}

