% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{gather_one}
\alias{gather_one}
\title{Gather one variable set from wide to long}
\usage{
gather_one(data, key_vars, varying, keep_others = FALSE, new_name = NULL,
  show_origin = FALSE)
}
\arguments{
\item{data}{The data frame that contains the variables}

\item{key_vars}{The key-variables or id variables that should be the grouping variables in the long format}

\item{varying}{The variables that vary and should be stacked in the long format}

\item{keep_others}{If \code{TRUE}, all other variables of the data frame are kept as well and stacked in the long format}

\item{new_name}{The new name of the stacked variable in the long format}

\item{show_origin}{Logical, if \code{TRUE}, and additional variable will be created showing from which variables the values originated. Default is to \code{FALSE}.}
}
\value{
Returns a data frame in long format
}
\description{
Gather one variable set from wide to long
}
\examples{
gather_one(data=wide_example_data, key_vars = "id", varying = c("var_p_1", "var_p_2"), new_name = "var")
}

