% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{generate_stems}
\alias{generate_stems}
\title{Generate unique item stems from item names}
\usage{
generate_stems(strings, search = "\\\\d+$")
}
\arguments{
\item{strings}{The item names that are used to generate the strings}

\item{search}{The unique element that identify a single item}
}
\value{
Returns the unique stems of a item list
}
\description{
Using this function and the \code{\link[padis]{select}} function, a lot of variables can easily be subsetted/selected from a data frame.
}
\examples{
names <- paste("var_", 1:5, sep = "")
generate_stems(names, search = "_\\\\d")
}

