% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_specific_functions.R
\name{get_selection_long}
\alias{get_selection_long}
\title{Turn Daily Data Into Issue Long Data}
\usage{
get_selection_long(data)
}
\arguments{
\item{data}{The daily data that should be turned into the long form data}
}
\value{
Returns the Selection data-set
}
\description{
This function is very specific and can only turn the daily data into the Selection long data.
}
\examples{
## not run
# these examples work only with the correct data frames
data_dy <- read.csv("Raw Diary Selection Data.csv", sep = ";", stringsAsFactors = FALSE)
issue <- get_issue_long(data_dy)
head(issue)
}

