% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_specific_functions.R
\name{match_ir}
\alias{match_ir}
\title{Match Two Items From Data Frame}
\usage{
match_ir(data, var_stem, end_pattern = c(".r", ".i"))
}
\arguments{
\item{data}{The data frame in which the variables should be found}

\item{var_stem}{The stem of the items, i.e. what both item-names share}

\item{end_pattern}{The pattern at the end of the variable that makes both variables unique}
}
\value{
A vector of length 2 with the two identified variables names
}
\description{
This function matches two items from a data frame \code{data} by looking for their common stem (\code{var_stem}), and the pattern at the end that makes each variable name unique (\code{end_pattern}). The function is mainly used in \code{\link[padis]{imaginer}}.
}
\examples{
data <- imaginer_example_data
match_ir(data, "PSP1.variable", c(".r", ".i"))
}

