% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_specific_functions.R
\name{other_day}
\alias{other_day}
\title{Transform Diary Data and Create a Next or Previous Day Variable}
\usage{
other_day(data, id = "PAR.ID", days_in = "Day.DaysIn",
  variables = c("Day.Wellbeing", "Day.Closeness"), prefix = NULL,
  next_day = FALSE)
}
\arguments{
\item{data}{The data frame that contains the variable}

\item{id}{The ID variable of the participants}

\item{days_in}{The variable identifying the day of each measurement. Must be numeric}

\item{variables}{The variables that should be transformed to a next day variable}

\item{prefix}{The prefix that is added to the names of the transformed variables. Default is NULL, which will set the prefix accoring to the \code{next_day} argument}

\item{next_day}{Logical. If \code{TRUE}, the next day is returned. If \code{FALSE}, the previous day is returned}
}
\value{
Returns a data frame with the original variables as well as the transformed variables
}
\description{
The function follows following rule: First, it treats each \code{id} separately. It then looks in the column \code{days_in} for each numeric value \code{x} that has a preceding or follwing lower or higher value \code{y} with a difference of \code{x - y = 1} (e.g. if there is a 4, it searches for a 3 or a 5, depending on the setting of \code{next_day}). If it finds a preceding value \code{y} that is one smaller than the initial value \code{x}, it returns corresponding row-wise values from the columns \code{variables} and stores them as in a new variable.
}
\examples{
data <- next_day_data
previous_day <- other_day(data)
next_day <- other_day(data, next_day = TRUE)
}

