% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_specific_functions.R
\name{real_imagined}
\alias{real_imagined}
\title{Combine .r and .i Pairs into a .c Variable}
\usage{
real_imagined(data, ignore_double = FALSE, combine_data = TRUE)
}
\arguments{
\item{data}{The data frame that contains the .r and .i variable pairs.}

\item{ignore_double}{logical. If set to TRUE, cases in which both .i and .r contain values will be ignored and NA will be defined to the .c variable. Default is to \code{FALSE}}

\item{combine_data}{logical. If set to TRUE, the original data frame will be combined with the new data frame. Default is to \code{TRUE}}
}
\value{
Returns a data frame with a combined variable .c, which contains the values of the .i and .r variables.
}
\description{
Combine .r and .i Pairs into a .c Variable
}
\examples{
real_imagined(real_imagined_data)
real_imagined(real_imagined_data, ignore_double = TRUE)
real_imagined(real_imagined_data, ignore_double = TRUE, combine_data=FALSE)
}

