% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{select_vars}
\alias{select_vars}
\title{Select variables from a data frame based on a prefix}
\usage{
select_vars(data, ids, prefix)
}
\arguments{
\item{data}{The data frame to look for the variables}

\item{ids}{The ID-variable(s) that should also be kept}

\item{prefix}{The prefix or pattern of the variables that are taken along}
}
\value{
It returns a data frame that contains only a subset of the previous data frame
}
\description{
Select variables from a data frame based on a prefix
}
\examples{
data <- mtcars
select_vars(data, "mpg", "^d.")
}

