% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{swap_char}
\alias{swap_char}
\title{Swaps character elements around strings}
\usage{
swap_char(string, separator = ".", unifyer = "_")
}
\arguments{
\item{string}{The string that should be swapped}

\item{separator}{The separator where the stringr should be splitted (takes always the first hit)}

\item{unifyer}{The character stringr with which the string should be merged again}
}
\value{
Returns a character stringr
}
\description{
This function is useful to generate the item names that are needed for the transformation from long to wide format
}
\examples{
swap_char("this.string", ".", "_")
}

