% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{write_to_wd}
\alias{write_to_wd}
\title{Write data frames directly to the working directory}
\usage{
write_to_wd(data, folder = NULL, type = "xlsx", name = NULL,
  overwrite = FALSE)
}
\arguments{
\item{data}{The data frame that should be saved}

\item{folder}{The folder in whoch the data is stored. If NULL, the current working directory is used.}

\item{type}{The type of the output, either \code{"xlsx"} or \code{"csv"}}

\item{name}{The name of the file. If NULL, the name of the object is used.}

\item{overwrite}{Logical, if \code{TRUE}, the file saved, even if it already exists. Default is to \code{FALSE}.}
}
\value{
The required output as a data file in the specified folder.
}
\description{
The function can be used to write data frames into the working directory (or any other specified path), either as .xlsx or .csv files.
}

