% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padis_general_functions.R
\name{write_to_ws}
\alias{write_to_ws}
\title{Write an object out of a function directly into the workspace}
\usage{
write_to_ws(object, overwrite = FALSE)
}
\arguments{
\item{object}{The object to be written into the workspace}

\item{overwrite}{Logical, if the object exists already in the workspace, it will be overwritten if set to \code{TRUE}}
}
\value{
Returns the object in the workspace.
}
\description{
This function is used to wirte objects from a function directly to the workspace or global environment.
}

